Add basic styling
Padding and margins will come later.

Added CSS class "recipe" to be used for recipe styling. The motivation
behind this is that I want all paragraph information about a recipe to
be in italics, but do not want this to interfere with any normal
non-recipe paragraphs that come along later.